+++ external/docker-integration-tests/src/test/scala/org/apache/spark/sql/jdbc/DockerJDBCIntegrationSuite.scala
@@ -66,7 +66,8 @@ abstract class DatabaseOnDocker {
   /**
    * Optional step before container starts
    */
-  def beforeContainerStart(hostConfigBuilder: HostConfig.Builder,
+  def beforeContainerStart(
+      hostConfigBuilder: HostConfig.Builder,
       containerConfigBuilder: ContainerConfig.Builder): Unit = {}
 }
